requirements.txt
discord.py==2.3.2
scikit-learn==1.2.2
tld==0.13
tldextract==5.0.1
ipwhois==1.2.0
python-whois==0.8.0
pandas==2.1.2
joblib==1.3.2
